refactor: move getPrecoUnitarioOuErro to produtos-service
