+++ src/services/produtos-service.js
@@ -1,5 +1,6 @@
 'use strict'
 
+const constant = require('../utils/constants')
 const Datastore = require('nedb-promises')
 const { join } = require('path')
 
@@ -39,6 +40,21 @@ exports.getPreco = async (produto) => {
 exports.criarProduto = async body => {
   body = formatarValores(body)
   return datastore.insert(body)
+}
+
+exports.getPrecoUnitarioOuErro = async (produto) => {
+  const quantidade = parseInt(produto.quantidade)
+  const { idProduto } = produto
+  const existePedido = await this.existeProduto({ _id: idProduto })
+  if (!existePedido) {
+    return { error: { statusCode: 400, message: constant.IDPRODUTO_INVALIDO, item: { idProduto, quantidade } } }
+  }
+
+  const { quantidade: quantidadeEstoque, preco } = await this.getDadosDoProduto({ _id: idProduto })
+  if (quantidade > quantidadeEstoque) {
+    return { error: { statusCode: 400, message: constant.ESTOQUE_INSUFICIENTE, item: { idProduto, quantidade, quantidadeEstoque } } }
+  }
+  return { precoUnitario: preco }
 }
 
 exports.deleteById = async id => {
